updated files and folders

+++ banking/src/main/java/com/banking/helper/AccountNumberGenerator.java
@@ -4,8 +4,8 @@ import java.util.Random;
 
 public class AccountNumberGenerator {
 	
-	public static int getAccountNumber() {
+	public static Long getAccountNumber() {
 	    String accountNo = String.format("%08d", new Random().nextInt(99999999));
-	    return Integer.parseInt(accountNo);
+	    return (long) Integer.parseInt(accountNo);
 	}
 }
